bugfixes on card edit view

+++ src/main/java/com/sbm4j/hearthstone/myhearthstone/model/DeckListItem.java
@@ -63,7 +63,7 @@ public class DeckListItem {
         this.heroCode = heroCode;
     }
 
-    public long getNbCards() {
+    public int getNbCards() {
         return nbCards;
     }
